editline: rework ec_editline_get_error_helps()
Currently, the tokenization is done in a quick and dirty way: when a space is
found in the string, it is replaced by a \0 and it retries to parse/complete to
generate the error helps.

Use the ec_strvec_sh_lex_str() function to enhance the tokenization. It makes
the function more complex, but it can be expected to be more reliable.

+++ include/ecoli/editline.h
@@ -285,7 +285,7 @@ void ec_editline_free_helps(struct ec_editline_help *helps, size_t n);
  */
 ssize_t ec_editline_get_error_helps(const struct ec_editline *editline,
 				    struct ec_editline_help **helps_out,
-				    int *char_idx);
+				    size_t *char_idx);
 
 /**
  * Set editline prompt.
